Models/EASE/requirements.txt
numpy==1.22.2
pandas==1.4.1
python-dateutil==2.8.2
pytz==2021.3
scipy==1.8.0
six==1.16.0
torch==1.10.2
tqdm==4.62.3
typing_extensions==4.1.1
python-dotenv==0.19.2
bottleneck==1.3.4
wandb==0.12.11